platform/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
jinja2>=3.1.2
pyyaml>=6.0.1
paramiko>=3.3.1
docker>=6.1.3
python-multipart>=0.0.6
